requirements.txt
asgiref
atomicwrites
attrs
certifi
charset-normalizer
colorama
Django
django-filter
djangorestframework
djangorestframework-simplejwt
docopt
drf-nested-routers
flake8
idna
iniconfig
mccabe
packaging
pipreqs
pluggy
py
pycodestyle
pyflakes
PyJWT
pytest
pytest-django
pytest-pythonpath
pytz
requests
sqlparse
toml
urllib3
yarg
